Simplified uv coordinates computation.

+++ source/Terrain.hpp
@@ -42,7 +42,7 @@ class Terrain
     const static int PATCH_RADIUS = 1;
     constexpr const static int PATCH_SIDE_LENGTH = 2 * PATCH_RADIUS + 1;
     constexpr const static int VERTICES_PER_SIDE = TerrainPatch::VERTICES_PER_SIDE * PATCH_SIDE_LENGTH;
-    constexpr const static int SIZE = static_cast<int>(TerrainPatch::SIZE) * PATCH_SIDE_LENGTH;
+    constexpr const static double SIZE = TerrainPatch::SIZE * static_cast<double>(PATCH_SIDE_LENGTH);
     const static int NO_GRASS_STRAW_PER_UNIT = 100;
     constexpr const static int NO_GRASS_STRAW = NO_GRASS_STRAW_PER_UNIT * SIZE * SIZE;
